+++ application/models/Units_model.php
@@ -140,6 +140,12 @@ class Units_model extends CI_Model {
         }
 	}
 	public function delete_unit($id){
+
+		if (demo_app()) {
+			echo "Demo không cho phép xóa";
+			return;
+		}
+
         $query1="delete from db_units where id=$id";
         if ($this->db->simple_query($query1)){
             echo "success";
